adding external functions
