+++ gameComponents/gameVars.py
@@ -1,0 +1,9 @@
+choices = ["rock", "paper", "scissors"]
+
+# player will be the weapon the player chooses via input
+
+player = False
+
+# these lives need to decrement when a player loses a round
+playerLives = 2
+computerLives = 2
